src/qumo/core/machine_key.py
"""Universal Key for Machine-Optimized Python Development.

This module serves as the canonical source of truth for machine optimization patterns
and principles across Python projects. It provides both documentation and practical
utilities for maintaining consistent machine-first code structure.
"""
from typing import Dict, Set, List, Tuple, Any, Optional
import re
import unicodedata

class MachineOptimizationKey:
    """KEY: Universal machine optimization patterns and principles."""
    PRINCIPLES = {'P1': 'Machine-First: Optimize for machine processing over human readability', 'P2': 'Pattern Consistency: Use uniform patterns across entire codebase', 'P3': 'Minimal Footprint: Reduce code size without losing functionality', 'P4': 'Self-Documenting: Embed meaning in patterns themselves', 'P5': 'Runtime Efficiency: Prioritize execution speed', 'P6': 'ML-Ready: Structure code for machine learning analysis', 'P7': 'Debuggable: Maintain error tracing capability', 'P8': 'Reversible: Keep all optimizations reversible', 'P9': 'Quantum-Ready: Support quantum computing patterns', 'P10': 'Unicode-Optimized: Utilize full Unicode space efficiently'}
    VARIABLE_PATTERNS: Dict[str, Dict[str, str]] = {'SELF': {'pattern': '\\b(self|this)\\b', 'symbol': 'ꜱ', 'purpose': 'Instance self-reference'}, 'PATH': {'pattern': '\\b(path|parent)\\b', 'symbol': 'ᴘ', 'purpose': 'File/directory paths'}, 'WINDOW': {'pattern': '\\b(window|widget)\\b', 'symbol': 'ᴡ', 'purpose': 'UI components'}, 'LOGGER': {'pattern': '\\b(logger|log)\\b', 'symbol': 'ʟ', 'purpose': 'Logging operations'}, 'DISPLAY': {'pattern': '\\b(display|directory)\\b', 'symbol': 'ᴅ', 'purpose': 'Display/directory operations'}, 'CURRENT': {'pattern': '\\b(current|command)\\b', 'symbol': 'ᴄ', 'purpose': 'Current state/command'}, 'ENGINE': {'pattern': '\\b(engine|event)\\b', 'symbol': 'ᴇ', 'purpose': 'Core engine/events'}, 'MANAGER': {'pattern': '\\b(manager|model)\\b', 'symbol': 'ᴍ', 'purpose': 'Management classes'}, 'VALUE': {'pattern': '\\b(value|view)\\b', 'symbol': 'ᴠ', 'purpose': 'Values/views'}, 'TEXT': {'pattern': '\\b(text|type)\\b', 'symbol': 'ᴛ', 'purpose': 'Text/type operations'}, 'FILE': {'pattern': '\\b(file|frame)\\b', 'symbol': 'ꜰ', 'purpose': 'File operations'}, 'ITEM': {'pattern': '\\b(item|image)\\b', 'symbol': 'ɪ', 'purpose': 'Items/images'}, 'OBJECT': {'pattern': '\\b(object|option)\\b', 'symbol': 'ᴏ', 'purpose': 'Objects/options'}, 'NAME': {'pattern': '\\b(name|node)\\b', 'symbol': 'ɴ', 'purpose': 'Names/nodes'}, 'QUEUE': {'pattern': '\\b(queue|query)\\b', 'symbol': 'ǫ', 'purpose': 'Queues/queries'}, 'RESULT': {'pattern': '\\b(result|response)\\b', 'symbol': 'ʀ', 'purpose': 'Results/responses'}}
    OPERATION_PATTERNS: Dict[str, Dict[str, str]] = {'WALRUS': {'pattern': ':=', 'purpose': 'Assignment with test', 'example': '(x := func())'}, 'CHAIN': {'pattern': ';', 'purpose': 'Operation chaining', 'example': 'x = 1; y = 2'}, 'FLOW': {'pattern': 'and/or', 'purpose': 'Flow control', 'example': 'x and y or z'}, 'FIRST': {'pattern': 'next()', 'purpose': 'First match', 'example': 'next((x for x in xs))'}, 'APPEND': {'pattern': '+=[x]', 'purpose': 'List append', 'example': 'lst += [item]'}, 'BOOL': {'pattern': '1/0', 'purpose': 'Boolean values', 'example': "'enabled': 1"}}
    METHOD_PATTERNS: Dict[str, Dict[str, str]] = {'GET': {'pattern': '\\bget_\\w+\\b', 'symbol': 'ɢ', 'purpose': 'Getters'}, 'SET': {'pattern': '\\bset_\\w+\\b', 'symbol': 'ꜱ', 'purpose': 'Setters'}, 'UPDATE': {'pattern': '\\bupdate_\\w+\\b', 'symbol': 'ᴜ', 'purpose': 'Updates'}, 'INIT': {'pattern': '\\binit_\\w+\\b', 'symbol': 'ɪ', 'purpose': 'Initialization'}, 'LOAD': {'pattern': '\\bload_\\w+\\b', 'symbol': 'ʟ', 'purpose': 'Loading'}, 'SAVE': {'pattern': '\\bsave_\\w+\\b', 'symbol': 'ꜱ', 'purpose': 'Saving'}}
    CLASS_PATTERNS: Dict[str, Dict[str, str]] = {'MANAGER': {'pattern': '\\w+Manager\\b', 'symbol': 'ᴍ', 'purpose': 'Management classes'}, 'CONTROLLER': {'pattern': '\\w+Controller\\b', 'symbol': 'ᴄ', 'purpose': 'Controller classes'}, 'SERVICE': {'pattern': '\\w+Service\\b', 'symbol': 'ꜱ', 'purpose': 'Service classes'}, 'MODEL': {'pattern': '\\w+Model\\b', 'symbol': 'ᴍ', 'purpose': 'Data models'}}
    RESERVED_WORDS: Set[str] = {'if', 'else', 'elif', 'while', 'for', 'in', 'is', 'not', 'and', 'or', 'True', 'False', 'None', 'try', 'except', 'finally', 'with', 'as', 'def', 'class', 'return', 'yield', 'break', 'continue', 'pass', 'raise', 'from', 'import', 'global', 'nonlocal', 'assert', 'del', 'lambda', 'len', 'str', 'int', 'float', 'list', 'dict', 'set', 'tuple', 'min', 'max', 'sum', 'any', 'all', 'zip', 'map', 'filter', 'i', 'j', 'k', 'x', 'y', 'z', 'n', 'm', '__init__', '__str__', '__repr__', '__len__', '__call__', '__enter__', '__exit__', '__iter__', '__next__', '__getitem__', '__setitem__', '__delitem__', '__contains__'}
    DOC_REQUIREMENTS: Dict[str, str] = {'KEY': 'Every optimized component must have a KEY comment', 'PATTERNS': 'Document all patterns used in the component', 'MAPPING': 'Maintain mapping between optimized and original forms', 'RATIONALE': 'Explain optimization choices in commit messages', 'IMPACT': 'Document performance/processing impact', 'REVERSIBILITY': 'Document how to reverse optimizations if needed'}
    QUANTUM_PATTERNS: Dict[str, Dict[str, str]] = {'SUPERPOSITION': {'pattern': '\\b(state|quantum)\\b', 'symbol': '⟨ψ⟩', 'purpose': 'Quantum state variables'}, 'ENTANGLEMENT': {'pattern': '\\b(linked|entangled)\\b', 'symbol': '⊗', 'purpose': 'Entangled operations'}, 'MEASUREMENT': {'pattern': '\\b(measure|observe)\\b', 'symbol': '⟨M⟩', 'purpose': 'State measurement'}, 'PROBABILITY': {'pattern': '\\b(probability|chance)\\b', 'symbol': '∫', 'purpose': 'Probability calculations'}, 'UNCERTAINTY': {'pattern': '\\b(uncertain|fuzzy)\\b', 'symbol': 'Δ', 'purpose': 'Uncertainty handling'}}
    UNICODE_PATTERNS: Dict[str, Dict[str, str]] = {'TENSOR': {'pattern': '\\b(tensor|matrix)\\b', 'symbol': '⊗', 'purpose': 'Tensor operations'}, 'VECTOR': {'pattern': '\\b(vector|array)\\b', 'symbol': '→', 'purpose': 'Vector operations'}, 'INFINITY': {'pattern': '\\b(infinite|unlimited)\\b', 'symbol': '∞', 'purpose': 'Infinite operations'}, 'PARTIAL': {'pattern': '\\b(partial|incomplete)\\b', 'symbol': '∂', 'purpose': 'Partial operations'}, 'SUMMATION': {'pattern': '\\b(sum|total)\\b', 'symbol': '∑', 'purpose': 'Summation operations'}, 'PRODUCT': {'pattern': '\\b(product|multiply)\\b', 'symbol': '∏', 'purpose': 'Product operations'}, 'INTEGRAL': {'pattern': '\\b(integral|continuous)\\b', 'symbol': '∫', 'purpose': 'Integration operations'}, 'EMPTY': {'pattern': '\\b(empty|void)\\b', 'symbol': '∅', 'purpose': 'Empty set/null operations'}, 'FORALL': {'pattern': '\\b(forall|foreach)\\b', 'symbol': '∀', 'purpose': 'Universal quantification'}, 'EXISTS': {'pattern': '\\b(exists|some)\\b', 'symbol': '∃', 'purpose': 'Existential quantification'}}
    MATH_PATTERNS: Dict[str, Dict[str, str]] = {'APPROX': {'pattern': '\\b(approximately|about)\\b', 'symbol': '≈', 'purpose': 'Approximate equality'}, 'NOTEQUAL': {'pattern': '\\b(not_equal|different)\\b', 'symbol': '≠', 'purpose': 'Inequality'}, 'SUBSET': {'pattern': '\\b(subset|contained)\\b', 'symbol': '⊆', 'purpose': 'Subset operations'}, 'INTERSECTION': {'pattern': '\\b(intersection|common)\\b', 'symbol': '∩', 'purpose': 'Set intersection'}, 'UNION': {'pattern': '\\b(union|combine)\\b', 'symbol': '∪', 'purpose': 'Set union'}}
    MARKERS: Dict[str, str] = {'OPTIMIZED': '⚡', 'QUANTUM': '⟨Q⟩', 'TENSOR': '⊗', 'VECTOR': '→', 'PARALLEL': '∥', 'ATOMIC': '⚛', 'ASYNC': '⟲', 'SYNC': '⟳'}
    UNICODE_BLOCKS = {'MATH_OPERATORS': (8704, 8959), 'MATH_SYMBOLS_A': (10176, 10223), 'MATH_SYMBOLS_B': (10624, 10751), 'MATH_OPERATORS_SUPP': (10752, 11007), 'MATH_ALPHANUMERIC': (119808, 120831), 'LATIN_EXTENDED_D': (42784, 43007)}

    @classmethod
    def get_pattern(cls, category: str, name: str) -> Optional[Dict[str, str]]:
        """Get pattern details by category and name."""
        patterns = getattr(cls, f'{category}_PATTERNS', {})
        return patterns.get(name)

    @classmethod
    def is_reserved(cls, word: str) -> bool:
        """Check if a word is reserved (should not be optimized)."""
        return word in cls.RESERVED_WORDS

    @classmethod
    def get_principle(cls, code: str) -> Optional[str]:
        """Get principle by code (P1-P8)."""
        return cls.PRINCIPLES.get(code)

    @classmethod
    def get_doc_requirement(cls, name: str) -> Optional[str]:
        """Get documentation requirement by name."""
        return cls.DOC_REQUIREMENTS.get(name)

    @classmethod
    def is_valid_unicode(cls, char: str) -> bool:
        """Check if a character is within allowed Unicode blocks."""
        if len(char) != 1:
            return False
        code_point = ord(char)
        for start, end in cls.UNICODE_BLOCKS.values():
            if start <= code_point <= end:
                return True
        allowed_categories = {'Sm', 'So', 'Ll', 'Lu'}
        category = unicodedata.category(char)
        if category not in allowed_categories:
            return False
        name = unicodedata.name(char, '').lower()
        prohibited = {'emoji', 'emoticon', 'dingbat', 'pictograph', 'ornamental'}
        return not any((p in name for p in prohibited))

    @classmethod
    def validate_unicode_usage(cls, text: str) -> List[str]:
        """Validate Unicode usage in text.
        
        Returns:
            List of validation messages for invalid characters.
        """
        messages = []
        for i, char in enumerate(text):
            if ord(char) > 127 and (not cls.is_valid_unicode(char)):
                name = unicodedata.name(char, 'UNKNOWN')
                messages.append(f'Invalid Unicode character at position {i}: {char} ({name}, U+{ord(char):04X})')
        return messages

    @classmethod
    def validate_optimization(cls, original: str, optimized: str) -> List[str]:
        """Validate optimization against principles and patterns."""
        messages = []
        if len(optimized) >= len(original):
            messages.append('P3: Optimization did not reduce code size')
        for category in ['VARIABLE', 'METHOD', 'CLASS']:
            patterns = getattr(cls, f'{category}_PATTERNS', {})
            for pattern_info in patterns.values():
                original_matches = len(re.findall(pattern_info['pattern'], original))
                symbol_matches = len(re.findall(pattern_info['symbol'], optimized))
                if original_matches != symbol_matches:
                    messages.append(f'P2: Inconsistent {category.lower()} pattern usage')
        for word in cls.RESERVED_WORDS:
            if word in original and word not in optimized:
                messages.append(f"Reserved word '{word}' was incorrectly optimized")
        messages.extend(cls.validate_unicode_usage(optimized))
        return messages

    @classmethod
    def get_quantum_pattern(cls, name: str) -> Optional[Dict[str, str]]:
        """Get quantum pattern details by name."""
        return cls.QUANTUM_PATTERNS.get(name)

    @classmethod
    def get_unicode_pattern(cls, name: str) -> Optional[Dict[str, str]]:
        """Get Unicode pattern details by name."""
        return cls.UNICODE_PATTERNS.get(name)

    @classmethod
    def get_math_pattern(cls, name: str) -> Optional[Dict[str, str]]:
        """Get mathematical pattern details by name."""
        return cls.MATH_PATTERNS.get(name)

    @classmethod
    def get_marker(cls, name: str) -> Optional[str]:
        """Get optimization marker by name."""
        return cls.MARKERS.get(name)

    @classmethod
    def get_fallback(cls, char: str) -> str:
        """Get ASCII fallback for a Unicode character."""
        fallbacks = {'⟨ψ⟩': 'psi', '⊗': 'x', '⟨M⟩': 'M', '∫': 'int', 'Δ': 'delta', '∑': 'sum', '∏': 'prod', '∞': 'inf', '∂': 'd', '∅': 'empty', '∀': 'forall', '∃': 'exists', '→': '->', '←': '<-', '⇒': '=>', '⇐': '<=', '∩': 'intersect', '∪': 'union', '⚛': 'atom', '∥': '||', '⟲': 'async', '⟳': 'sync'}
        return fallbacks.get(char, char)